+++ test/sqlforcql/config.clj
@@ -5,6 +5,6 @@
 (defn db-test-fixture
   "Connects to the default db, uses the default session and then disconnects from the db."
   [f]
-  (core/connect-to-default-db "sqlforcql")
+  (core/connect-to-default-db "localhost" "cassandra" "cassandra" "syndeia_cloud_store")
   (f)
   (core/disconnect-from-default-db))