+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## 0.3.1 - 2021-05-14
+### Added
+- Fix insert statement generation bugs:
+  - generate a boolean value as true instead of 'true'
+  - generate an empty hash set value as {\'\'} instead of {\"\"}
+  - generate a string value containing single-quotes with two single quotes i.e. Part(''ab'') instead of Part('ab')
+
 ## 0.3.0 - 2021-05-11
 ### Added
 - `generate` namespace containing functions `get-insert-statements` & `get-insert-statement`.